// frontend/src/api/generated/model/ndmsInterfaceInventoryResponseRequiredGuardsItem.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * keen-pbr API
 * REST API for the keen-pbr policy-based routing daemon.
 * OpenAPI spec version: 3.0.0
 */

export type NdmsInterfaceInventoryResponseRequiredGuardsItem = typeof NdmsInterfaceInventoryResponseRequiredGuardsItem[keyof typeof NdmsInterfaceInventoryResponseRequiredGuardsItem];


export const NdmsInterfaceInventoryResponseRequiredGuardsItem = {
  typed_rci: 'typed_rci',
  automatic_backup: 'automatic_backup',
  ownership_check: 'ownership_check',
  optimistic_revision: 'optimistic_revision',
} as const;